winit-core: move application
