+++ winit-core/src/lib.rs
@@ -3,6 +3,7 @@ pub mod as_any;
 pub mod cursor;
 #[macro_use]
 pub mod error;
+pub mod application;
 pub mod event;
 pub mod event_loop;
 pub mod icon;
